requirements.txt
aiohttp==3.7.4
dateparser==1.1.8
dblpy==0.4.0
discord.py==2.2.2
motor==2.3.0
pytz==2023.3
regex==2023.3.23
requests==2.24.0
Unidecode==1.1.1
websockets==9.1
matplotlib==3.7.1
multidict==4.5.0
pymongo==3.11.0
numpy==1.22
six==1.16

